+++ server/src/routes/index.ts
@@ -2,6 +2,7 @@ import { Router } from 'express';
 import { healthRouter } from '../modules/health/health.route.js';
 import { authRouter } from '../modules/auth/auth.route.js';
 import { userRouter } from '../modules/user/user.route.js';
+import { incidentRouter } from '../modules/incident/incident.route.js';
 
 const apiRouter: Router = Router();
 
@@ -13,5 +14,8 @@ apiRouter.use('/auth', authRouter);
 
 // Mount User Routes
 apiRouter.use('/users', userRouter);
+
+// Mount Incident Routes
+apiRouter.use('/incidents', incidentRouter);
 
 export { apiRouter };